+++ commerce/warehouse/src/main/java/ru/yandex/practicum/commerce/warehouse/entity/OrderBooking.java
@@ -23,5 +23,5 @@ public class OrderBooking {
     @CollectionTable(name="order_products", joinColumns = @JoinColumn(name = "order_id"))
     @MapKeyColumn(name = "product_id")
     @Column(name = "quantity")
-    private Map<UUID, Long> products;
+    private Map<UUID, Integer> products;
 }
